Enable Selection for Determining TK
Allow the user to specify the TK_det parameter for choosing whether to take a maximum, minimum or mean on the value of TK computed in the computeWettingAndDrying subroutine. TK_det is an integer where 1 = max, 2 = min, 3 = mean, and the default value is 0, which means that TK is not chosen deterministically. This also means that the default behavior is somewhat stochastic, in that solutions may depend on the ordering of computations, e.g., the order of elements in the connectivity table.