Update Strategy for Friction (TK) in Wet/Dry
Instead of saving either the mean, max or min of the depth-averaged friction, TK, (based on choice of TK_det) now only save the max if TK_det = 1. This avoids setting TK too low if it has already been set to a higher value by the nodal attributes module, e.g., in the Apply2DBottomFriction routine.