Scale Factor and Add Offset for NetCDF Met Inputs
Some NetCDF meterological input files for ADCIRC use a scale factor and add offset to allow the storing the variables as integers in the input files, whereby the values shouled then be converted to real values by using the scale factor and add offset. This code allows ADCIRC to read in the factor and offset and apply them to the variable values read in. Unfortunately, this code was also based on an older version of ADCIRC, so many of these codes revert recent changes done in ADCIRC. I plan to reconcile these differences myself eventually, but for now I am storing the "old" code here in this commit for posterity, since it is known to work.